@@ -15,17 +15,17 @@ class ApiProblemTest extends TestCase
 {
     public function statusCodes()
     {
-        return array(
-            '200' => array(200),
-            '201' => array(201),
-            '300' => array(300),
-            '301' => array(301),
-            '302' => array(302),
-            '400' => array(400),
-            '401' => array(401),
-            '404' => array(404),
-            '500' => array(500),
-        );
+        return [
+            '200' => [200],
+            '201' => [201],
+            '300' => [300],
+            '301' => [301],
+            '302' => [302],
+            '400' => [400],
+            '401' => [401],
+            '404' => [404],
+            '500' => [500],
+        ];
     }
 
     /**
@@ -86,13 +86,13 @@ class ApiProblemTest extends TestCase
         $payload    = $apiProblem->toArray();
         $this->assertArrayHasKey('exception_stack', $payload);
         $this->assertInternalType('array', $payload['exception_stack']);
-        $expected = array(
-            array(
+        $expected = [
+            [
                 'code' => $exceptionChild->getCode(),
                 'message' => $exceptionChild->getMessage(),
                 'trace' => $exceptionChild->getTrace(),
-            ),
-        );
+            ],
+        ];
         $this->assertEquals($expected, $payload['exception_stack']);
     }
 
@@ -106,12 +106,12 @@ class ApiProblemTest extends TestCase
 
     public function knownStatusCodes()
     {
-        return array(
-            '404' => array(404),
-            '409' => array(409),
-            '422' => array(422),
-            '500' => array(500),
-        );
+        return [
+            '404' => [404],
+            '409' => [409],
+            '422' => [422],
+            '500' => [500],
+        ];
     }
 
     /**
@@ -153,7 +153,7 @@ class ApiProblemTest extends TestCase
             'Invalid input',
             'http://example.com/api/problem/400',
             'Invalid entity',
-            array('foo' => 'bar')
+            ['foo' => 'bar']
         );
         $this->assertEquals('bar', $problem->foo);
     }
@@ -165,7 +165,7 @@ class ApiProblemTest extends TestCase
             'Invalid input',
             'http://example.com/api/problem/400',
             'Invalid entity',
-            array('foo' => 'bar')
+            ['foo' => 'bar']
         );
         $array   = $problem->toArray();
         $this->assertArrayHasKey('foo', $array);
@@ -179,7 +179,7 @@ class ApiProblemTest extends TestCase
             'Invalid input',
             'http://example.com/api/problem/400',
             'Invalid entity',
-            array('title' => 'SHOULD NOT GET THIS')
+            ['title' => 'SHOULD NOT GET THIS']
         );
         $array   = $problem->toArray();
         $this->assertArrayHasKey('title', $array);
@@ -209,7 +209,7 @@ class ApiProblemTest extends TestCase
     public function testUsesAdditionalDetailsFromExceptionWhenProvided()
     {
         $exception  = new Exception\DomainException('exception message', 401);
-        $exception->setAdditionalDetails(array('foo' => 'bar'));
+        $exception->setAdditionalDetails(['foo' => 'bar']);
         $apiProblem = new ApiProblem('401', $exception);
         $payload    = $apiProblem->toArray();
         $this->assertArrayHasKey('foo', $payload);
@@ -218,13 +218,13 @@ class ApiProblemTest extends TestCase
 
     public function invalidStatusCodes()
     {
-        return array(
-            '-1'  => array(-1),
-            '0'   => array(0),
-            '7'   => array(7),  // reported
-            '14'  => array(14), // observed
-            '600' => array(600),
-        );
+        return [
+            '-1'  => [-1],
+            '0'   => [0],
+            '7'   => [7],  // reported
+            '14'  => [14], // observed
+            '600' => [600],
+        ];
     }
 
     /**
